requirements.txt
keyboard==0.13.5
Pillow>=8.3.1,<9.0.0

